refactor: EmojiService의 로직 책임 분리
1. SvgDocument를 생성하는 로직 SvgDocumentFactory로 분리
2. EmojiRandomIndexGenerator에서 index의 범위 부분도 documentFactory로 이동
3. RandomNumberGenerator를 인터페이스로 분리
