+++ src/main/java/clickme/clickme/util/EmojiRandomIndexGenerator.java
@@ -1,16 +1,16 @@
 package clickme.clickme.util;
 
+import org.springframework.stereotype.Component;
+
 import java.util.Random;
 
-public class EmojiRandomIndexGenerator {
+@Component
+public class EmojiRandomIndexGenerator implements RandomNumberGenerator {
 
-    private static final int RANDOM_BOUND = 15;
     private static final Random random = new Random();
 
-    private EmojiRandomIndexGenerator() {
-    }
-
-    public static int getRandomNumber() {
-        return random.nextInt(RANDOM_BOUND);
+    @Override
+    public int generator(int size) {
+        return random.nextInt(size);
     }
 }
