+++ src/main/java/clickme/clickme/util/RandomNumberGenerator.java
@@ -1,0 +1,6 @@
+package clickme.clickme.util;
+
+public interface RandomNumberGenerator {
+
+    int generator(int size);
+}
